fix: api 구조 수정
좀 더 REST API 에 가깝도록 구조 수정
